requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=9.0.0
albumentations>=1.3.0

# Deep Learning frameworks
timm>=0.9.0
transformers>=4.30.0
accelerate>=0.20.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.1.0

# Configuration and logging
PyYAML>=6.0
hydra-core>=1.3.0
tensorboard>=2.12.0
wandb>=0.15.0

# Image processing
imageio>=2.25.0
matplotlib>=3.6.0
seaborn>=0.11.0

# Utilities
tqdm>=4.64.0
rich>=13.0.0
click>=8.1.0
pathlib2>=2.3.0

# Optional: SAM2 dependencies (if using SAM2 teacher)
# segment-anything2>=1.0.0
# hydra-core>=1.3.0

# Optional: Mamba dependencies (if using Mamba backbone)
# mamba-ssm>=1.0.0
# causal-conv1d>=1.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0